requirements.txt
psycopg2-binary
natsort
python-dotenv